refactor(config): remove dead code and stale comments, unify helpers
Dead code removed:
- parser/lexer.rs (tokenize() returned an empty vec; no callers)
- config-side types::OutboundIndex (unused; runtime uses the
  honk-ebpf-common type of the same name)
- Node::to_uri, RoutingCondition::display_expr,
  Default for DnsRequestAction, DnsRouting::has_legacy_only,
  the unused ExperimentalConfig re-export, decode_url_fragment wrapper

Comment fixes:
- dns.rs: rewrite the stale serde(skip) note (manual impls, not an
  attribute), drop section banners (project convention)
- share_link.rs: ECH bare-ech comment now reflects HTTPS-RR discovery
  being implemented; drop the inaccurate 'API server import paths' claim
- node.rs: Group doc lists all four policies; drop name-restating field
  docs across node/subscription/routing/dns

Style unification:
- default_true() and parse_duration_secs() now live once in types.rs
  (were defined 3x / 2x with diverging semantics)

AGENTS.md updated to match.
