+++ AGENTS.md
@@ -65,14 +65,14 @@ Configuration schema and parsers used by the rest of the project. Deps are pure-
 
 - `src/config.rs` — top-level `Config` / `GlobalConfig` (~40 global fields), `from_file` / `to_file` / `validate`, JSON helpers, and `ensure_builtin_nodes()` (injects the built-in `direct` node, mapped to `DirectHandler` via `NodeProtocol::HTTP`, idempotent). **The crate never calls `ensure_builtin_nodes()` itself** — `honk-core` calls it at startup and SIGHUP reload (`honk-core/src/lib.rs`); other consumers must call it explicitly.
 - Format loading: the dae syntax is primary; TOML/YAML/JSON serde loaders remain for compatibility (undocumented). `from_file` picks by extension — recognized `.json`/`.yaml`/`.toml` try that format then fall back only among TOML/YAML/JSON (dae is never tried); unknown/missing extensions use the file-aware dae loader (including `include`) then try TOML → YAML → JSON.
-- `src/parser/` — the dae-syntax parser. **`parser/mod.rs` is the entire real parser** (line/section based, ~1600 lines): file-aware `include` expansion (quoted/bare paths, glob, nested entry-relative resolution, canonical directory boundary, duplicate/cycle rejection), all sections, group policies and aliases (`fixed(0)`/`select`→Selector, `min_moving_avg`/`min_avg10`/`min_last_delay`→URLTest, `roundrobin`→LoadBalance, `fallback`→Fallback), nested `filter: group('a','b')` (comma or pipe separated) routed into `Group.groups`, DNS `upstream`/`routing { request/response }`, `fixed_domain_ttl`, `subscription`, `experimental`, `resolve_group_filters`. `parser/lexer.rs` is **dead scaffolding** (`tokenize()` returns an empty vec); `section_parser.rs` is only the `Section` struct.
+- `src/parser/` — the dae-syntax parser. **`parser/mod.rs` is the entire real parser** (line/section based, ~1600 lines): file-aware `include` expansion (quoted/bare paths, glob, nested entry-relative resolution, canonical directory boundary, duplicate/cycle rejection), all sections, group policies and aliases (`fixed(0)`/`select`→Selector, `min_moving_avg`/`min_avg10`/`min_last_delay`→URLTest, `roundrobin`→LoadBalance, `fallback`→Fallback), nested `filter: group('a','b')` (comma or pipe separated) routed into `Group.groups`, DNS `upstream`/`routing { request/response }`, `fixed_domain_ttl`, `subscription`, `experimental`, `resolve_group_filters`. `section_parser.rs` is only the `Section` struct.
 - Group filter resolution: `resolve_group_filters` resolves only node filters (`name('exact')`, `name(keyword: 'pat')` — case-sensitive substring); the "include all nodes" fallback applies **only** when a group has neither node filters nor sub-groups.
 - `src/node.rs` — `Node` (all per-protocol fields, incl. **ECH**: `ech_enabled` / `ech_config` / `ech_config_path`) **plus `Group` and `GroupPolicy`** (Selector/URLTest/LoadBalance/Fallback). `src/group.rs` is a 2-line re-export. `Group.groups: Vec<String>` holds nested sub-group tags; `Group.default`, `final_outbound` (dae `final:`), `tolerance` (default 50 ms), `idle_timeout`, `interrupt_connections`.
 - `src/dns.rs` — much richer than a plain upstream list: `DnsConfig` (`upstream`, `routing`, `strategy`, `cache`, `fixed_domain_ttl`), `DnsUpstream` (name, address, `protocol: DnsProtocol`, `tls_server_name`, **`outbound: Option<String>`** — per-upstream dial-path proxy tag), and the dae-shaped DNS routing model (`DnsRequestRule`/`DnsResponseRule` with AND-ed `DnsCond`s — Qname/Qtype/Upstream/Ip, each negatable — first match wins; actions Reject/AsIs/Accept/Upstream(name); legacy `rules`/`fallback` with conversion). `types.rs::DnsProtocol` has 6 variants: Udp, Tcp, Tls (DoT), Https (DoH), H3 (DoH3), Quic (DoQ). Request/response routing types are populated only by the dae parser (deliberately outside the serde tree).
 - `src/share_link.rs` — `Node::from_share_link`, the single share-link parser: SIP002 `ss://` (all base64 forms, plugin suffix), `ssr://` base64 blobs, `vmess://` base64-JSON (v2rayN schema), trojan/trojan-go ws/grpc query params, AnyTLS pool params, hysteria2 auth/obfs/bandwidth (`upmbps`/`downmbps`)/port-hopping (`mport`/`mhop`)/QUIC-window params and `pinSHA256=<hex>` cert pins, **ECH params** (`ech_config=<base64url ECHConfigList>`, `ech=1`), plus socks5/http(s)/vless/hysteria2/tuic/juicity. Node name = decoded `#fragment`, else `scheme-host` (never leaks credentials). Chain links (`a -> b`): only the first hop is parsed. Used by the dae parser and `honk-core` subscriptions.
 - `src/routing.rs` — `RoutingRule` (condition + outbound + priority + **`must` flag** — Go dae semantics: match does not finalize, continues searching), `RoutingCondition` (14 matcher lists incl. dscp, ip_version, mac, process_name, geoip/geosite), `RoutingOutbound` (Simple/Complex), `RoutingConfig`.
 - `src/experimental.rs` — `ExperimentalConfig` { `clash_api: ClashApiConfig` (`external_controller`, `external_ui`, `secret`, **`default_mode`**, default "Rule"), `cache_file: CacheFileConfig` (`enabled`, `path` default `cache.db`, `cache_id`, `store_fakeip`, `store_dns`) } — also parsed from the dae `experimental { ... }` section.
-- `src/subscription.rs`, `src/types.rs` (`NodeProtocol` 12 variants, `DialMode` ip/domain/domain+/domain++, config-side `OutboundIndex` — **a different type** from `honk-ebpf-common`'s `#[repr(u8)] OutboundIndex`, `SubscriptionType`, `DnsProtocol`), `src/error.rs` (`ConfigError`).
+- `src/subscription.rs`, `src/types.rs` (`NodeProtocol` 12 variants, `DialMode` ip/domain/domain+/domain++, `SubscriptionType`, `DnsProtocol`, plus the shared `default_true`/`parse_duration_secs` helpers), `src/error.rs` (`ConfigError`).
 
 ### `crates/honk-ebpf-common`
